Increase the member update timeout
Discord Gateway (through discordgo) will sleep for a number of seconds (seems discord responds with 10 seconds) and attempt the request one more time after that deadline.  Role-srv is not waiting around because it's got shit to get done.  Currently there doesn't seem to be a way within discordgo to disable this reattempt so giving role-srv some patience seems like the best approach for now.